# packages/webui/frontend/src/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "linling-webui",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_api_agents_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSummary"
                  },
                  "title": "Response List Agents Api Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{name}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent",
        "operationId": "get_agent_api_agents__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{name}/triggers": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Triggers",
        "description": "Return the agent's bot's matchable DSL triggers for the inline-suggest panel.\n\nCheap to call: a provider closure walks the live classifier each\ntime so hot-reload reflects on the very next poll. When the agent\nisn't backed by a bot (test harness, agent-only deployments) we\nreturn an empty list — the composer panel hides itself in that case.\n\nReused across both panels: the chat composer's autocomplete and any\nfuture ``/help`` mirror in the SPA. RBAC: same as ``/api/agents`` —\nvisible to every authenticated role; triggers don't reveal anything\na user couldn't discover by typing.",
        "operationId": "list_triggers_api_agents__name__triggers_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TriggerSuggestion"
                  },
                  "title": "Response List Triggers Api Agents  Name  Triggers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{name}/memory": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Memory",
        "description": "Return the short-term sliding window for (user, scope).\n\nLong-term memory (vector store) is not yet exposed through the runtime;\nreturned as empty until Task 18.2 lands.",
        "operationId": "memory_api_agents__name__memory_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "webui",
              "title": "User Id"
            }
          },
          {
            "name": "scope_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "test",
              "title": "Scope Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_MemoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{name}/chat": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Chat",
        "description": "Single-shot chat. Routes through the DSL classifier first.\n\nMirrors ``/ws/agents/{name}/stream``: when ``attach_bot_to_webui``\nhas installed a web chat dispatcher for this agent, we delegate\nso triggers like ``我的灵玉`` reach their ``.ling`` handler. Without\na dispatcher (e.g. minimal test harness) we fall back to a direct\n``runtime.invoke`` like before.",
        "operationId": "chat_api_agents__name__chat_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "token_today": {
            "type": "integer",
            "title": "Token Today",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider",
          "model"
        ],
        "title": "AgentSummary"
      },
...
      "ChatRequest": {
        "properties": {
          "input": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Input"
          },
          "history_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Key"
          },
          "scope_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tool_calls_made": {
            "type": "integer",
            "title": "Tool Calls Made",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "agent"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ChatSegment"
            },
            "type": "array",
            "title": "Segments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "content",
          "latency_ms"
        ],
        "title": "ChatResponse"
      },
      "ChatSegment": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "alt": {
            "type": "string",
            "title": "Alt",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ChatSegment",
        "description": "One piece of a rich chat reply (text or image).\n\nMirrors the WS frame shape so the SPA can use the same renderer\non both transports."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TriggerSuggestion": {
        "properties": {
          "raw": {
            "type": "string",
            "title": "Raw"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "has_args": {
            "type": "boolean",
            "title": "Has Args",
            "default": false
          },
          "literal_prefix": {
            "type": "string",
            "title": "Literal Prefix",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "raw",
          "label"
        ],
        "title": "TriggerSuggestion",
        "description": "One DSL trigger surfaced to the chat composer's inline-suggest panel.\n\nThe frontend uses ``label`` for display, ``literal_prefix`` for the\npre-fill on click (so triggers with regex placeholders like ``(.*)``\ndon't paste a still-not-matchable shape into the input), and\n``has_args`` to decide whether a click should auto-send or just\npark the cursor at the placeholder."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "_MemoryView": {
        "properties": {
          "short_term": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Short Term"
          },
          "long_term": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Long Term",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "short_term"
        ],
        "title": "_MemoryView"
      },
...
    }
  }
}